backend/requirements.txt
asgiref==3.7.2
black==23.10.0
click==8.1.7
Django==4.2.6
django-filter==23.3
django-model-utils==4.3.1
django-pandas==0.6.6
djangorestframework==3.14.0
flake8==6.1.0
isort==5.12.0
Markdown==3.5
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
platformdirs==3.11.0
psycopg==3.1.12
psycopg-binary==3.1.12
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
